tests

+++ hospitalModules/hospital/src/main/java/hospital/dao/DoctorDAO.java
@@ -2,10 +2,11 @@ package hospital.dao;
 
 import hospital.model.Staff;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface DoctorDAO {
     public List<Staff> getAllDoctors();
     public Staff getDoctorBySurname(String surname);
-    Staff getById(int id);
+    Staff getById(BigInteger id);
 }
