+++ hospitalModules/hospital/src/main/java/hospital/dao/PrescriptionDAO.java
@@ -2,13 +2,14 @@ package hospital.dao;
 
 import hospital.model.Prescription;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface PrescriptionDAO {
 
-    Prescription getPrescriptionById(int id);
+    Prescription getPrescriptionById(BigInteger id);
 
-    List<Prescription> getAllPrescriptions(int id);
+    List<Prescription> getAllPrescriptions(BigInteger id);
 
     void updatePrescription(Prescription presc);
