+++ hospitalModules/hospital/src/main/java/hospital/model/DiagnosisType.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 
 import javax.persistence.*;
+import java.math.BigInteger;
 import java.util.List;
 
 @Entity
@@ -12,7 +13,7 @@ public class DiagnosisType {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name="id_diagnosis_type")
-    private Integer id;
+    private BigInteger id;
     @Column
     private String title;
     @OneToMany(mappedBy = "diagnosisType", cascade = CascadeType.MERGE, orphanRemoval = true,fetch = FetchType.EAGER)
@@ -54,11 +55,11 @@ public class DiagnosisType {
         this.diagnosises = diagnosises;
     }
 
-    public Integer getId() {
+    public BigInteger getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(BigInteger id) {
         this.id = id;
     }
