+++ hospitalModules/hospital/src/main/java/hospital/model/Event.java
@@ -3,6 +3,7 @@ package hospital.model;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 import javax.persistence.*;
+import java.math.BigInteger;
 import java.util.Calendar;
 import java.util.Date;
 
@@ -14,7 +15,7 @@ public class Event {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name="id_event")
-    private Integer id;
+    private BigInteger id;
     @ManyToOne(cascade = CascadeType.ALL,fetch = FetchType.EAGER)
     @JoinColumn(name = "id_patient")
     private Patient patient;
@@ -54,11 +55,11 @@ public class Event {
         this.procedure = procedure;
     }
 
-    public Integer getId() {
+    public BigInteger getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(BigInteger id) {
         this.id = id;
     }
